get_dom_node_map_entry(): take name as a dom_string
Requires that dom_string->length becomes a size_t.

+++ src/document/dom/string.h
@@ -2,7 +2,7 @@
 #define EL__DOCUMENT_DOM_STRING_H
 
 struct dom_string {
-	uint16_t length;
+	size_t length;
 	unsigned char *string;
 };
